fix(deprecations): limit to single entry
