--- src/utility/validation/deprecation.js
@@ -1,34 +1,0 @@
-/*
- * Copyright 2020 American Express Travel Related Services Company, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
-// internal functions to formalize API deprecation
-
-export function createDeprecationMessage(message = '') {
-  return [
-    '[One Service Worker]: Deprecation Notice - %s is marked for deprecation and will not be accessible in the next major release.',
-    message,
-  ]
-    .join('\n')
-    .trim();
-}
-
-export function deprecationNotice(deprecatedMember, message) {
-  // assumes any deprecated items to be functions for the time being
-  return function warnOnDeprecation(...args) {
-    console.warn(createDeprecationMessage(message), deprecatedMember.name);
-    return deprecatedMember(...args);
-  };
-}
